Lower the number of activities, closes #2
