+++ app/src/main/AndroidManifest.xml
@@ -54,21 +54,6 @@
 
         <activity
             android:name=".Activity5"
-            android:label="@string/title_activity5" >
-        </activity>
-
-        <activity
-            android:name=".Activity6"
-            android:label="@string/title_activity6" >
-        </activity>
-
-        <activity
-            android:name=".Activity7"
-            android:label="@string/title_activity7" >
-        </activity>
-
-        <activity
-            android:name=".Activity8"
             android:label="@string/title_activity8" >
         </activity>
     </application>
