+++ app/src/main/res/values/strings.xml
@@ -20,16 +20,13 @@
     <string name="maternal_grand_mother">الجدة لأم:</string>
     <string name="paternal_grand_father">الجدة لأب:</string>
     <string name="next">التالي</string>
-    <string name="title_activity2">الزوج(ة) والأولاد</string>
     <string name="wives">الزوجات:</string>
     <string name="husband">الزوج:</string>
     <string name="daughters">البنات:</string>
     <string name="sons">الأبناء:</string>
-    <string name="title_activity3">الأحفاد والإخوة لأم</string>
-    <string name="title_activity4">الإخوة</string>
-    <string name="title_activity5">أبناء الإخوة</string>
-    <string name="title_activity6">الأعمام</string>
-    <string name="title_activity7">أبناء الأعمام</string>
+    <string name="title_activity2">الأبناء والأحفاد</string>
+    <string name="title_activity3">الإخوة</string>
+    <string name="title_activity4">أبناء الإخوة، الأعمام وأبناء الأعمام</string>
     <string name="title_activity8">حل المسألة</string>
 
 </resources>
